Move dynamic gas pricing logic into own EIP files
